Normalize config before initialization
